Fixed compile errors and updated project files for windows build. (Merged from 1.12RC)
git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@6936 18c085ea-50e0-402c-830e-de6fd14e8384

+++ src/avt/Filters/avtStreamlineWrapper.C
@@ -155,7 +155,7 @@ avtStreamlineWrapper::Debug()
     sl->PtEnd(end);
     debug1<<"******seed: "<<end;
     debug1<<" Dom= [ "<<domain<<", ";
-    for (int i = 0; i < seedPtDomainList.size(); i++)
+    for (size_t i = 0; i < seedPtDomainList.size(); i++)
         debug1<<seedPtDomainList[i]<<", ";
     debug1<<"] ";
     debug1<< " steps= "<<sl->size()<<endl;
@@ -307,7 +307,7 @@ avtStreamlineWrapper::ComputeStatistics()
     maxCnt = 0;
     sum = 0;
     numDomainsVisited = 0;
-    for (int i = 0; i < domainVisitCnts.size(); i++)
+    for (size_t i = 0; i < domainVisitCnts.size(); i++)
     {
         int cnt = domainVisitCnts[i];
         if (cnt > maxCnt)
